# Speedup reading of large DXF documents