I'm testing out the capabilities of netDxf in a project I'm working on. I've noticed that loading large documents takes a pretty long time. After taking the code through the performance analyzer, I noticed two low hanging fruit optimizations.

The first one is to switch the internal list of EntityCollection to a HashSet. Most of the CPU time for loading the file was spent in the 'Contains' method of the internal list. This means there is no index based retrieving of entities anymore though. HashSet is much more  performant in checking if an object is already contained in the collection. You might want to keep both a HashSet and a List internally if index based retrieval is a must. There would be some syncing logic but I think it will still be way faster.

This change reduced the loading time of a 5MB DXF file from 13.9 seconds to 2.5 seconds.

The seconds one is in the function of IsValidName. The fastest way I could find was to use the IndexOfAny method. This further reduces the loading time of my file from 2.5 to 1.8 seconds.

Feel free to make or suggest changes if you've unhappy with some part of the PR.