@haplokuon 2019-04-29
Thank you for pointing this out. The process of adding an entity to a document has become painfully slow by one simple and completely avoidable mistake. Checking if an entity is contained inside the list is not needed. One of the reasons I added the Owner property for all DxfObjects is to know if the object already belongs to to somebody else, entities cannot be shared between blocks, once an entity is added to a block an owner is assigned, therefore checking for that is enough, there is no need to use the Contains method. I will make the necessary arrangements for this and I look for other Contains calls that might have slipped through.

The problem with using a HashSet is, for what I know form the documentation, that it cannot guarantee the order of the elements. Usually it doesn't matter the order of the entities inside the list but sometimes it does, for example, when drawing lines over an image; or when working with colors you might want to have one over another.

About, your proposal on the check for invalid characters, I am perfectly fine with it, if it is faster welcome be. If you isolate this change I will merge it, or I can manually added myself, your choice.

Daniel