@haplokuon 2019-06-23
Added, in netDxf 2.3.0, your suggested improvement for the InvalidCharacters method of the TableObject class.